phone number at receipt

+++ src/main/java/com/greentechpay/bff/dto/response/ResponseDto.java
@@ -3,11 +3,9 @@ package com.greentechpay.bff.dto.response;
 import lombok.Builder;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 @Builder
 public class ResponseDto<T> {
-    private BigDecimal totalAmount;
+    private Double totalAmount;
     private T data;
 }
